Reorganize web TS sources into the target layout
Renames script.ts to main.ts and groups view-related files under
web/src/view/ in preparation for introducing Viewport, Renderer, and
Layer abstractions. Pure file-move commit: no behavior changes.

Also generalizes the front-end .gitignore globs so they keep working as
new compiled paths get added.
