+++ .gitignore
@@ -49,9 +49,5 @@ bin/
 ### Node / TypeScript front-end ###
 web/node_modules/
 # Compiled TypeScript output served by Ktor — rebuild with `npm run build`.
-src/main/resources/web/script.js
-src/main/resources/web/script.js.map
-src/main/resources/web/types.js
-src/main/resources/web/types.js.map
-src/main/resources/web/constants/
-src/main/resources/web/rendering/
+src/main/resources/web/**/*.js
+src/main/resources/web/**/*.js.map
