+++ src/main/resources/web/index.html
@@ -19,6 +19,6 @@
 </head>
 <body>
 <canvas id="simCanvas"></canvas>
-<script src="script.js" type="module"></script>
+<script src="main.js" type="module"></script>
 </body>
 </html>
