+++ web/src/view/coordinates-constants.ts
